Extract the parser into a separate usable class.
